LPD-23303 Create useLoadingObserver
